requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aniso8601==8.0.0
astor==0.7.1
attrs==20.2.0
bottle==0.12.18
bottle-websocket==0.2.9
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
dataclasses==0.7
decorator==4.4.2
Eel==0.12.2
fastapi==0.61.1
firebase-admin==4.4.0
Flask==1.1.2
Flask-API==2.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-Testing==0.8.0
future==0.18.2
gast==0.3.3
gevent==1.4.0
gevent-websocket==0.10.1
google-api-core==1.22.4
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
greenlet==0.4.15
grpcio==1.32.0
h11==0.9.0
h5py==2.10.0
httplib2==0.18.1
idna==2.8
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2
MarkupSafe==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
msgpack==1.0.0
numpy==1.18.1
Pillow==7.0.0
proglog==0.1.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydantic==1.6.1
pyreadline==2.1
pyrsistent==0.17.3
python-multipart==0.0.5
pytz==2020.1
requests==2.22.0
rsa==4.6
scipy==1.3.1
six==1.14.0
starlette==0.13.6
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.49.0
uritemplate==3.0.1
urllib3==1.25.8
uvicorn==0.11.8
websockets==8.1
Werkzeug==1.0.0
whichcraft==0.6.1
wincertstore==0.2
wrapt==1.11.2
zipp==3.1.0
